Create a Ratio Trait

A ratio trait is the ratio of total abundance of two groups of glycans. For example, the ratio of complex glycans and hybrid glycans, or the ratio of bisecting and unbisecting glycans.

Usage

ratio(num_cond, denom_cond, within = NULL, na_action = "keep")

Arguments

num_cond

Condition to use for defining the numerator. An expression that evaluates to a logical vector. The names of all built-in meta-properties (see all_mp_fns()) and custom meta-properties can be used in the expression.

denom_cond

Condition to use for defining the denominator. Same format as num_cond.

within

Condition to set a restriction for the glycans. Same format as num_cond.

na_action

How to handle missing values.

"keep" (default): keep the missing values as NA.

"zero": set the missing values to 0.

Value

A derived trait function.

Details

You can use ratio() to create ratio trait easily.

For example:

# Ratio of complex glycans and hybrid glycans
ratio(T == "complex", T == "hybrid")

# Ratio of bisecting and unbisecting glycans
ratio(B, !B)

# Ratio of core-fucosylated and non-core-fucosylated glycans within complex glycans
ratio(nFc > 0 & T == "complex", nFc == 0 & T == "complex")

# The above example can be simplified as:
ratio(nFc > 0, nFc == 0, within = (T == "complex"))  # more readable

Note that the last example uses & for logical AND. Actually, you can use any logical operator in the expression in R (e.g., |, !, etc.).

prop() is a special case of ratio(), i.e., prop(cond, within) is equivalent to ratio(cond & within, within). We recommend using prop() instead of ratio() for clarity if possible.