Calculate Derived Traits

This function calculates derived traits from glycomic or glycoproteomic profiles. For glycomics data, it calculates the derived traits directly. For glycoproteomics data, each glycosite is treated as a separate glycome, and derived traits are calculated in a site-specific manner.

derive_traits(): Calculate derived traits from a glyexp::experiment() object.

derive_traits_(): Calculate derived traits from a tibble in tidy format. Use this function if you are not using the glyexp package.

Usage

derive_traits(exp, trait_fns = NULL, mp_fns = NULL)

derive_traits_(tbl, data_type, trait_fns = NULL, mp_fns = NULL)

Arguments

exp

A glyexp::experiment() object. Before using this function, you should preprocess the data using the glyclean package. For glycoproteomics data, the data should be aggregated to the "gfs" (glycoforms with structures) level using glyclean::aggregate(). Also, please make sure that the glycan_structure column is present in the var_info table, as not all glycoproteomics identification softwares provide this information. "glycan_structure" can be a glyrepr::glycan_structure() vector, or a character vector of glycan structure strings supported by glyparse::auto_parse().

trait_fns

A named list of derived trait functions created by trait factories. Names of the list are the names of the derived traits. Default is NULL, which means all derived traits in all_traits() are calculated.

mp_fns

A named list of meta-property functions. This parameter is useful if your trait functions use custom meta-properties other than those in all_mp_fns(). Default is NULL, which means all meta-properties in all_mp_fns() are used.

tbl

A tibble in tidy format, with the following columns:

sample: sample ID

glycan_structure: glycan structures, either a glyrepr::glycan_structure() vector or a character vector of glycan structure strings supported by glyparse::auto_parse().

value: the quantification of the glycan in the sample.

For glycoproteomics data, additional columns are needed:

protein: protein ID

protein_site: the glycosite position on the protein The unique combination of protein and protein_site determines a glycosite.

Other columns are ignored.

Please make sure that the data has been properly preprocessed, including normalization, missing value handling, etc. Specifically, for glycoproteomics data, please make sure that the data has been aggregated to the "glycoforms with structures" level. That is the quantification of each glycan structure on each glycosite in each sample.

data_type

Either "glycomics" or "glycoproteomics". Only needed for derive_traits_().

Value

For derive_traits(): a new glyexp::experiment() object for derived traits with "traitomics" type. Instead of "quantification of each glycan on each glycosite in each sample", the new experiment() contains "the value of each derived trait on each glycosite in each sample", with the following columns in the var_info table:

variable: variable ID

trait: derived trait name

For glycoproteomics data, with additional columns:

protein: protein ID

protein_site: the glycosite position on the protein

Other columns in the var_info table (e.g. gene) are retained if they have "many-to-one" relationship with glycosites (unique combinations of protein, protein_site). That is, each glycosite cannot have multiple values for these columns. gene is a common example, as a glycosite can only be associate with one gene. Descriptions about glycans are not such a column, as a glycosite can have multiple glycans, thus having multiple descriptions. Columns not having this relationship with glycosites will be dropped. Don't worry if you cannot understand this logic, as long as you know that this function will try its best to preserve useful information.

sample_info and meta_data are not modified, except that the exp_type field of meta_data is set to "traitomics".

For derive_traits_(): a tidy tibble containing the following columns:

sample: sample ID

trait: derived trait name

value: the value of the derived trait

For glycoproteomics data, with additional columns:

protein: protein ID

protein_site: the glycosite position on the protein

Other columns in the original tibble are not included.